Add cmdio.Command interface
Due to interface assertions not working on embedded interface types,
code that inspected the capabilities of an underlying command
represented as an io.ReadWriter failed, requiring manual passthrough of
all implemented methods.

This was not a sustainable way of handling methods with differing
sets of command implementations.

cmdio.Command represents the broadest possible interface of any
command. For implementations that choose not to implement optional
interfaces, NopCommand is provided as an embeddable type to
automatically fulfill the interface with dummy implementations, which
can be selectively overridden as desired.